test: add Excel color parser tests
